wgpu: Moved Draw creation outside of backend
